+++ apps/cowswap-frontend/src/common/pure/CancelButton/index.tsx
@@ -1,4 +1,4 @@
-import { PropsWithChildren } from 'react'
+import { PropsWithChildren, ReactNode } from 'react'
 
 import { Command } from '@cowprotocol/types'
 import { UI, LinkStyledButton } from '@cowprotocol/ui'
@@ -10,9 +10,7 @@ export type CancelButtonProps = {
   className?: string
 } & PropsWithChildren
 
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export function CancelButton({ onClick, children, className }: CancelButtonProps) {
+export function CancelButton({ onClick, children, className }: CancelButtonProps): ReactNode {
   return (
     <LinkStyledButton onClick={onClick} className={className} color={`var(${UI.COLOR_DANGER})`}>
       {children || <Trans>Cancel order</Trans>}