+++ libs/types/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [3.3.0](https://github.com/cowprotocol/cowswap/compare/types-v3.2.0...types-v3.3.0) (2026-02-20)
+
+
+### ✨ Features
+
+* **chains:** disable Lens and support deprecating networks ([#7028](https://github.com/cowprotocol/cowswap/issues/7028)) ([c12d6c2](https://github.com/cowprotocol/cowswap/commit/c12d6c21ad69258c587378add9da86d72ff755b2))
+
 ## [3.2.0](https://github.com/cowprotocol/cowswap/compare/types-v3.1.0...types-v3.2.0) (2026-02-13)